Interesting refactor
