+++ geode-redis/src/main/java/org/apache/geode/redis/internal/executor/string/MSetNXExecutor.java
@@ -19,12 +19,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.geode.cache.Region;
-import org.apache.geode.redis.internal.ByteArrayWrapper;
-import org.apache.geode.redis.internal.Coder;
-import org.apache.geode.redis.internal.Command;
-import org.apache.geode.redis.internal.ExecutionHandlerContext;
+import org.apache.geode.redis.internal.*;
 import org.apache.geode.redis.internal.RedisConstants.ArityDef;
-import org.apache.geode.redis.internal.RedisDataTypeMismatchException;
 
 public class MSetNXExecutor extends StringExecutor {
 
@@ -50,7 +46,7 @@ public class MSetNXExecutor extends StringExecutor {
       byte[] keyArray = commandElems.get(i);
       ByteArrayWrapper key = new ByteArrayWrapper(keyArray);
       try {
-        checkDataType(key, context);
+        checkDataType(key, RedisDataType.REDIS_STRING, context);
       } catch (RedisDataTypeMismatchException e) {
         hasEntry = true;
         break;
